+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* Automatically remove empty sections
+### Fixed
+* CLI actually write the changes to disk :)
 
 ## [1.0.0] - 2017-06-20
 ### Added